+++ libqiprog/src/qiprog_usb_device.c
@@ -116,9 +116,10 @@ void qiprog_change_device(struct qiprog_device *new_dev)
 	/* Initialize the new driver */
 	qi_dev->drv->dev_open(qi_dev);
 }
+
 static uint8_t ctrl_buf[64];
 
-qiprog_err qiprog_handle_control_request(uint8_t bRequest, uint8_t wValue,
+qiprog_err qiprog_handle_control_request(uint8_t bRequest, uint16_t wValue,
 					 uint16_t wIndex, uint16_t wLength,
 					 uint8_t ** data, uint16_t * len)
 {
